refactor: Change environment.py to keyed.py and change to abs imports
